feat(dev): add autogen and OS trust of SSL cert
**DO NOT MERGE!!! This PR is dependent on #13.**

ServiceWorker and WebSocket requests will increasingly require not only
SSL to be enabled, but also for the self-signed SSL certificate
typically used in development to be trusted by the developer's browser.
(When the certificate is not trusted, the browser regards the HTTPS site
as "Not Secure", even when it's working.)

Past solutions usually involve undocumented browser flags [reference:
https://www.chromium.org/blink/serviceworker/service-worker-faq](https://www.chromium.org/blink/serviceworker/service-worker-faq)
and don't work for all browsers. A better solution is to use the
operating system's mechanisms to trust self-signed certificates, so the
development environment most closely resembles a production environment.

This addition adds a `before` hook to the `webpack-dev-server`
configuration. It attempts to create an OSX-compatible self-signed
certificate, store it where `webpack-dev-server` can use it, and export
it to the OSX Keychain with trusted status.

To preserve security, the dev server also attempts to untrust and remove
this certificate before it closes or quits. Furthermore, the
certificates themelves are only valid for 30 days if they happen to
linger anyway.

However, a password is still required, and sometimes a password for
exiting, which is a little weird.

Added
=====

 - necessary dependencies for script
 - lib directory for miscellaneous build scripting, separate from the
 CRA-produced special "scripts" dir
 - implementation of OSX SSL trust as a Node script
 - webpack.config.js platform detection switch and before hook

Known Issues
============

 - Password prompting is a little weird. Asking for a sudo password
 before dev server starts is unexpected, but understandable. (Sudo cache
 intervals could also be expanded here.)
 - Only works for OSX presently, and invokes a lot of perhaps-volatile
 internals (i.e. the `security` command). We know that Windows
 PowerShell can be used to programmatically install and trust a
 self-signed cert, so I expect that to be a later addition.

 **DO NOT MERGE!!! This PR is dependent on #13.**